refactor: Change MinIO config to JUnit extension
