+++ dhis-2/dhis-support/dhis-support-hibernate/src/main/java/org/hisp/dhis/dbms/HibernateDbmsManager.java
@@ -27,14 +27,11 @@
  */
 package org.hisp.dhis.dbms;
 
-import java.util.ArrayList;
-import java.util.List;
 import javax.persistence.EntityManager;
 import lombok.extern.slf4j.Slf4j;
 import org.hisp.dhis.cache.HibernateCacheManager;
 import org.springframework.jdbc.BadSqlGrammarException;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Component;
 
 /**
@@ -382,48 +379,6 @@ public class HibernateDbmsManager implements DbmsManager {
     } catch (BadSqlGrammarException ex) {
       log.debug("Table " + table + " does not exist");
     }
-  }
-
-  @Override
-  public boolean tableExists(String tableName) {
-    final String sql =
-        "select table_name from information_schema.tables "
-            + "where table_name = '"
-            + tableName
-            + "' "
-            + "and table_type = 'BASE TABLE'";
-
-    List<Object> tables = jdbcTemplate.queryForList(sql, Object.class);
-
-    return tables != null && tables.size() > 0;
-  }
-
-  @Override
-  public List<List<Object>> getTableContent(String table) {
-    List<List<Object>> tableContent = new ArrayList<>();
-
-    SqlRowSet sqlRowSet = jdbcTemplate.queryForRowSet("select * from " + table);
-    int cols = sqlRowSet.getMetaData().getColumnCount() + 1;
-
-    List<Object> headers = new ArrayList<>();
-
-    for (int i = 1; i < cols; i++) {
-      headers.add(sqlRowSet.getMetaData().getColumnName(i));
-    }
-
-    tableContent.add(headers);
-
-    while (sqlRowSet.next()) {
-      List<Object> row = new ArrayList<>();
-
-      for (int i = 1; i < cols; i++) {
-        row.add(sqlRowSet.getObject(i));
-      }
-
-      tableContent.add(row);
-    }
-
-    return tableContent;
   }
 
   // -------------------------------------------------------------------------
