+++ dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/app/AppManagerTest.java
@@ -34,13 +34,16 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import java.io.IOException;
+// import org.hisp.dhis.app.AppManagerTest.MinIOConfig;
 import org.hisp.dhis.appmanager.App;
 import org.hisp.dhis.appmanager.AppManager;
 import org.hisp.dhis.appmanager.AppStatus;
-import org.hisp.dhis.test.config.MinIOTestConfig;
 import org.hisp.dhis.test.integration.PostgresIntegrationTestBase;
+import org.hisp.dhis.test.junit.MinIOTestExtension;
+import org.hisp.dhis.test.junit.MinIOTestExtension.MinIOConfig;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.FileSystemResource;
@@ -52,7 +55,8 @@ import org.springframework.test.context.ContextConfiguration;
  * Test class configured for use cases when DHIS2 is configured to use MinIO storage. The default
  * storage is the local filesystem.
  */
-@ContextConfiguration(classes = {MinIOTestConfig.class})
+@ExtendWith(MinIOTestExtension.class)
+@ContextConfiguration(classes = {MinIOConfig.class})
 class AppManagerTest extends PostgresIntegrationTestBase {
 
   @Autowired AppManager appManager;
